add: 第五章、第六章

+++ EMHC_CLASS/ConsoleApp1/class/EMHC_CONSULTANT.cs
@@ -48,7 +48,7 @@ namespace ConsoleApp1
         /// <summary>
         /// 人次
         /// </summary>
-        public int? NUMBER { get; set; }
+        public decimal? NUMBER { get; set; }
 
         /// <summary>
         /// 建立人員
